+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add gracefulness when using the new data source health check endpoint.
   Apparently, this is not implemented thoroughly for all data source types yet.
 * Add new factory methods `GrafanaApi.{from_url(),from_env()}`.
+* Add `GrafanaApi.connect()` and `GrafanaApi.version()`.
 
 
 ## 2.3.0 (2022-05-26)